Bowling kata started
Initial tests, and a Frame class that will handle each individual frame.

Current thinking is that each frame will go into the Bowling class, and
that will handle multiple frames and calculate the score. Only problem
is that the last frame is a "special case" and can have more than two
throws.

Also, need to decide how actual bowls will be inputted. Perhaps just an
array of pins knocked down, and the bowling class can "chunk" them (so
two throws are one frame, unless the first throw was a strike)