Tests for double spare and double strike
Spare works beautifully. Strike breaks my whole current calculation
method. I had assumed (ass!) that even with strikes you only needed to
know the score of the NEXT frame to be able to calculate yours for
example...

    Frame 1    | Frame 2    | Frame 3
    Strike     | Strike     | 5

I thought that the score for frame 1 would be 20 (strike + strike) and
the score for frame 2 would be 15 (strike + 5). However, the damn thing
is recursive ... so the score for Frame 1 is actually 25 (strike + frame
2 (aka: 15))

This may well require _yet another_ rethink of how I work this out. On
the plus side, the tests are pretty good :P