MMmm.... Magic! Working OK :)
I spent quite some time trying to figure out a good way to write the
logic for calculating the final frame, and it wasn't until I'd written
out a load of tests that I realised that they were doing a simple sum.

I feel like a giant idiot, but it was nice validation for the tests.
This explains why I've go so many for the end game.